+++ src/lib/Core/CHANGELOG.md
@@ -1,4 +1,7 @@
-﻿# 3.1.0
+﻿# 3.1.1
+* More robust nucleotide detection.
+
+# 3.1.0
 * Updated density data model.
 * Removed the support for the DSN6 density data format.
 * Fixed a bug in marching cubes that sometimes added bad triangles because of a vertex index cache not being properly cleaned.